Revert "Refactor server architecture: consolidate Server, SSEManager, and FileWatcher into separate classes for improved organization and maintainability"
This reverts commit 0c298e8eb73526605837870fc984703c4170ebc2.
